Created bookPreview.js web component
Created `book-preview` Web Component to render book preview details with encapsulated styling in order to avoid conflict as well as removing the original css styling. The component uses the shadow DOM to isolate and dynamically render book information. Introduced a `bookData` setter to pass the book details, which automatically triggers the `render()` function, making the component modular and reusable.